requirements.txt
fastapi
uvicorn
python-multipart
moviepy
Pillow
numpy
pydantic
aiofiles
jinja2
edge-tts
